Level_3/Fake_News_Detection/requirements.txt
beautifulsoup4==4.14.2
groq==0.32.0
markdownify==1.2.0
python-dotenv==1.1.1
Requests==2.32.5
streamlit==1.50.0
